src/domain/forum/application/use-cases/questions/fetch-question-comments/use-case.ts
import { success, type Either } from '@/core/either-failure-or-success'
import { QuestionCommentsRepositoryInterface } from '@/domain/forum/application/repositories'
import type { CommentWithAuthor } from '@/domain/forum/enterprise/entities/value-objects/comment-with-author'
import { Injectable } from '@nestjs/common'

type FetchQuestionCommentsUseCaseRequest = {
  page: number
  questionId: string
}

type FetchQuestionCommentsUseCaseResponse = Either<null, { comments: CommentWithAuthor[] }>

@Injectable()
export class FetchQuestionCommentsUseCase {
  constructor(private questionCommentsRepository: QuestionCommentsRepositoryInterface) {}

  async execute({
    page,
    questionId,
  }: FetchQuestionCommentsUseCaseRequest): Promise<FetchQuestionCommentsUseCaseResponse> {
    const comments = await this.questionCommentsRepository.findManyByQuestionIdWithAuthor(
      questionId,
      { page }
    )

    return success({
      comments,
    })
  }
}
